refactor(server): move CFProcessor to recceiver/cf, delete cfstore.py
