+++ server/recceiver/cf/__init__.py
@@ -1,0 +1,3 @@
+from recceiver.cf.processor import CFProcessor
+
+__all__ = ["CFProcessor"]
